move responseHandler to middleware folder
